Rename constants.ts to interfaces.ts
