+++ packages/react-core/src/modules/slack/index.ts
@@ -1,3 +1,3 @@
 export * from "./context";
 export * from "./hooks";
-export * from "./constants";
+export * from "./interfaces";
