+++ packages/react-core/src/modules/slack/interfaces.ts
